Provide state mixins for shared state: Position and Color
Move states into a States folder.
