+++ Sources/Interaction/Widgets2/States/CubeState/index.js
@@ -1,25 +1,20 @@
 import macro from 'vtk.js/Sources/macro';
 import vtkWidgetState from 'vtk.js/Sources/Interaction/Widgets2/WidgetState';
+import vtkPositionMixin from 'vtk.js/Sources/Interaction/Widgets2/States/PositionMixin';
+import vtkColorMixin from 'vtk.js/Sources/Interaction/Widgets2/States/ColorMixin';
 
 // ----------------------------------------------------------------------------
 
-function vtkSphereState(publicAPI, model) {
-  model.classHierarchy.push('vtkSphereState');
-
-  // --------------------------------------------------------------------------
-
-  publicAPI.translate = (dx, dy, dz) => {
-    const [x, y, z] = publicAPI.getPositionByReference();
-    publicAPI.setPosition(x + dx, y + dy, z + dz);
-  };
+function vtkCubeState(publicAPI, model) {
+  model.classHierarchy.push('vtkCubeState');
 }
 
 // ----------------------------------------------------------------------------
 
 const DEFAULT_VALUES = {
-  position: [0, 0, 0],
-  radius: 1,
-  color: 0.5,
+  xLength: 1,
+  yLength: 1,
+  zLength: 1,
 };
 
 // ----------------------------------------------------------------------------
@@ -28,15 +23,17 @@ export function extend(publicAPI, model, initialValues = {}) {
   Object.assign(model, DEFAULT_VALUES, initialValues);
 
   vtkWidgetState.extend(publicAPI, model);
-  macro.setGet(publicAPI, model, ['radius', 'color']);
-  macro.setGetArray(publicAPI, model, ['position'], 3);
+  vtkPositionMixin.extend(publicAPI, model);
+  vtkColorMixin.extend(publicAPI, model);
 
-  vtkSphereState(publicAPI, model);
+  macro.setGet(publicAPI, model, ['radius', 'xLength', 'yLength', 'zLength']);
+
+  vtkCubeState(publicAPI, model);
 }
 
 // ----------------------------------------------------------------------------
 
-export const newInstance = macro.newInstance(extend, 'vtkSphereState');
+export const newInstance = macro.newInstance(extend, 'vtkCubeState');
 
 // ----------------------------------------------------------------------------
