+++ Sources/Interaction/Widgets2/States/PositionMixin/index.js
@@ -1,10 +1,9 @@
 import macro from 'vtk.js/Sources/macro';
-import vtkWidgetState from 'vtk.js/Sources/Interaction/Widgets2/WidgetState';
 
 // ----------------------------------------------------------------------------
 
-function vtkCubeState(publicAPI, model) {
-  model.classHierarchy.push('vtkCubeState');
+function vtkPositionMixin(publicAPI, model) {
+  model.classHierarchy.push('vtkPositionMixin');
 
   // --------------------------------------------------------------------------
 
@@ -18,10 +17,6 @@ function vtkCubeState(publicAPI, model) {
 
 const DEFAULT_VALUES = {
   position: [0, 0, 0],
-  xLength: 1,
-  yLength: 1,
-  zLength: 1,
-  color: 0.5,
 };
 
 // ----------------------------------------------------------------------------
@@ -29,23 +24,11 @@ const DEFAULT_VALUES = {
 export function extend(publicAPI, model, initialValues = {}) {
   Object.assign(model, DEFAULT_VALUES, initialValues);
 
-  vtkWidgetState.extend(publicAPI, model);
-  macro.setGet(publicAPI, model, [
-    'radius',
-    'color',
-    'xLength',
-    'yLength',
-    'zLength',
-  ]);
   macro.setGetArray(publicAPI, model, ['position'], 3);
 
-  vtkCubeState(publicAPI, model);
+  vtkPositionMixin(publicAPI, model);
 }
 
 // ----------------------------------------------------------------------------
 
-export const newInstance = macro.newInstance(extend, 'vtkCubeState');
-
-// ----------------------------------------------------------------------------
-
-export default { newInstance, extend };
+export default { extend };
