+++ Sources/Interaction/Widgets2/States/SphereState/example/controlPanel.html
